+++ pkg/models/query.go
@@ -16,6 +16,9 @@ const (
 	QueryTypeListAssociatedAssets = "ListAssociatedAssets"
 	QueryTypeDescribeAsset        = "DescribeAsset"
 	QueryTypeDescribeAssetModel   = "DescribeAssetModel"
+	QueryTypeListAssetProperties  = "ListAssetProperties"
+	QueryTypeListTimeSeries       = "ListTimeSeries"
+	QueryTypeExecuteQuery         = "ExecuteQuery"
 )
 
 const (
@@ -28,20 +31,25 @@ const (
 )
 
 type BaseQuery struct {
+	// General
 	AwsRegion string `json:"region,omitempty"`
-	// Deprecated: use assetIds
-	AssetId             string   `json:"assetId,omitempty"`
-	AssetIds            []string `json:"assetIds,omitempty"`
-	PropertyId          string   `json:"propertyId,omitempty"`
-	PropertyAlias       string   `json:"propertyAlias,omitempty"`
-	NextToken           string   `json:"nextToken,omitempty"`
-	MaxPageAggregations int      `json:"maxPageAggregations,omitempty"`
-	ResponseFormat      string   `json:"responseFormat,omitempty"`
+	QueryType string `json:"-"`
 
+	// Sitewise specific
+	// Deprecated: use assetIds
+	AssetId             string            `json:"assetId,omitempty"`
+	AssetIds            []string          `json:"assetIds,omitempty"`
+	PropertyId          string            `json:"propertyId,omitempty"`
+	PropertyAlias       string            `json:"propertyAlias,omitempty"`
+	NextToken           string            `json:"nextToken,omitempty"`
+	NextTokens          map[string]string `json:"nextTokens,omitempty"`
+	MaxPageAggregations int               `json:"maxPageAggregations,omitempty"`
+	ResponseFormat      string            `json:"responseFormat,omitempty"`
+
+	// Also provided by sqlutil.Query. Migrate to that
 	Interval      time.Duration     `json:"-"`
 	TimeRange     backend.TimeRange `json:"-"`
 	MaxDataPoints int64             `json:"-"`
-	QueryType     string            `json:"-"`
 }
 
 // MigrateAssetId handles AssetId <--> AssetIds backward compatibility.