+++ src/components/browser/AssetBrowser.tsx
@@ -1,15 +1,16 @@
 import React, { Component } from 'react';
 import { Button, Icon, Modal, Spinner, Tab, TabContent, TabsBar } from '@grafana/ui';
 import { AssetInfo } from '../../types';
-import { DataSource } from 'DataSource';
+import { DataSource } from 'SitewiseDataSource';
 import { SitewiseCache } from 'sitewiseCache';
 import { BrowseModels } from './BrowseModels';
 import { BrowseHierarchy } from './BrowseHierarchy';
+import { type Region } from '../../regions';
 
 export interface Props {
   datasource: DataSource;
   assetId?: string; // The incoming value
-  region?: string;
+  region?: Region;
   onAssetChanged: (assetId?: string) => void;
 }
 
@@ -40,7 +41,7 @@ export class AssetBrowser extends Component<Props, State> {
   }
 
   async componentDidUpdate(oldProps: Props) {
-    let update: State = { ...this.state };
+    const update: State = { ...this.state };
     let shouldUpdate = false;
 
     if (this.props.region !== oldProps.region) {