+++ src/types.ts
@@ -1,16 +1,22 @@
-import { DataQuery, SelectableValue } from '@grafana/data';
-import { AwsAuthDataSourceJsonData, AwsAuthDataSourceSecureJsonData } from '@grafana/aws-sdk';
+import type { AwsAuthDataSourceJsonData, AwsAuthDataSourceSecureJsonData } from '@grafana/aws-sdk';
+import type { SelectableValue } from '@grafana/data';
+import { QueryEditorMode } from '@grafana/plugin-ui';
+import type { DataQuery } from '@grafana/schema';
+import type { Region } from './regions';
 
 // Matches https://github.com/grafana/iot-sitewise-datasource/blob/main/pkg/models/query.go#L3
 export enum QueryType {
   ListAssetModels = 'ListAssetModels',
   ListAssets = 'ListAssets',
   ListAssociatedAssets = 'ListAssociatedAssets',
+  ListAssetProperties = 'ListAssetProperties',
   DescribeAsset = 'DescribeAsset',
   PropertyValue = 'PropertyValue',
   PropertyValueHistory = 'PropertyValueHistory',
   PropertyAggregate = 'PropertyAggregate',
   PropertyInterpolated = 'PropertyInterpolated',
+  ListTimeSeries = 'ListTimeSeries',
+  ExecuteQuery = 'ExecuteQuery',
 }
 
 export enum SiteWiseQuality {
@@ -36,6 +42,7 @@ export enum SiteWiseResolution {
   TenSec = '10s',
   Min = '1m',
   TenMin = '10m',
+  FifteenMin = '15m',
   Hour = '1h',
   TenHour = '10h',
   Day = '1d',
@@ -52,8 +59,14 @@ export enum AggregateType {
 
 export interface SitewiseQuery extends DataQuery {
   queryType: QueryType;
-  region?: string; // aws region string
+  region?: Region; // aws region string
   responseFormat?: SiteWiseResponseFormat;
+
+  // QueryEditor
+  editorMode?: QueryEditorMode;
+
+  // RawQueryEditor
+  rawSQL?: string;
 
   /** @deprecated -- this is migrated to assetIds */
   assetId?: string;
@@ -64,7 +77,9 @@ export interface SitewiseQuery extends DataQuery {
   quality?: SiteWiseQuality;
   resolution?: SiteWiseResolution;
   lastObservation?: boolean;
+  flattenL4e?: boolean;
   maxPageAggregations?: number;
+  clientCache?: boolean;
 }
 
 export interface SitewiseNextQuery extends SitewiseQuery {
@@ -73,6 +88,7 @@ export interface SitewiseNextQuery extends SitewiseQuery {
    * will require multiple pages in order to fulfil the requests
    */
   nextToken?: string;
+  nextTokens?: Record<string, string>;
 }
 
 /**
@@ -80,10 +96,6 @@ export interface SitewiseNextQuery extends SitewiseQuery {
  */
 export interface ListAssetModelsQuery extends SitewiseQuery {
   queryType: QueryType.ListAssetModels;
-}
-
-export function isListAssetModelsQuery(q?: SitewiseQuery): q is ListAssetModelsQuery {
-  return q?.queryType === QueryType.ListAssetModels;
 }
 
 /**
@@ -103,22 +115,12 @@ export function isListAssetsQuery(q?: SitewiseQuery): q is ListAssetsQuery {
 
 export interface ListAssociatedAssetsQuery extends SitewiseQuery {
   queryType: QueryType.ListAssociatedAssets;
-  hierarchyId?: string; // if empty, will list the parents
+  loadAllChildren?: boolean; // When passed, we will loop through all associated hierarchies, and return children from all.
+  hierarchyId?: string; // if empty and loadAllChildren is false, will list the parents
 }
 
 export function isListAssociatedAssetsQuery(q?: SitewiseQuery): q is ListAssociatedAssetsQuery {
   return q?.queryType === QueryType.ListAssociatedAssets;
-}
-
-/**
- * {@link https://docs.aws.amazon.com/iot-sitewise/latest/APIReference/API_ListAssetModels.html}
- */
-export interface DescribeAssetQuery extends SitewiseQuery {
-  queryType: QueryType.DescribeAsset;
-}
-
-export function isDescribeAssetQuery(q?: SitewiseQuery): q is ListAssetModelsQuery {
-  return q?.queryType === QueryType.DescribeAsset;
 }
 
 /**
@@ -127,10 +129,8 @@ export function isDescribeAssetQuery(q?: SitewiseQuery): q is ListAssetModelsQue
  */
 export interface AssetPropertyValueQuery extends SitewiseQuery {
   queryType: QueryType.PropertyValue;
-}
 
-export function isAssetPropertyValueQuery(q?: SitewiseQuery): q is AssetPropertyValueQuery {
-  return q?.queryType === QueryType.PropertyAggregate;
+  flattenL4e?: boolean;
 }
 
 /**
@@ -140,6 +140,7 @@ export interface AssetPropertyValueHistoryQuery extends SitewiseQuery {
   queryType: QueryType.PropertyValueHistory;
 
   timeOrdering?: SiteWiseTimeOrder;
+  flattenL4e?: boolean;
 }
 
 export function isAssetPropertyValueHistoryQuery(q?: SitewiseQuery): q is AssetPropertyValueHistoryQuery {
@@ -167,11 +168,21 @@ export function isAssetPropertyAggregatesQuery(q?: SitewiseQuery): q is AssetPro
  */
 export interface AssetPropertyInterpolatedQuery extends SitewiseQuery {
   queryType: QueryType.PropertyInterpolated;
-  timeOrdering?: SiteWiseTimeOrder;
 }
 
 export function isAssetPropertyInterpolatedQuery(q?: SitewiseQuery): q is AssetPropertyInterpolatedQuery {
   return q?.queryType === QueryType.PropertyInterpolated;
+}
+
+/**
+ * {@link https://docs.aws.amazon.com/iot-sitewise/latest/APIReference/API_ListTimeSeries.html}
+ */
+
+export interface ListTimeSeriesQuery extends SitewiseQuery {
+  queryType: QueryType.ListTimeSeries;
+  aliasPrefix?: string;
+  assetId?: string;
+  timeSeriesType?: 'ASSOCIATED' | 'DISASSOCIATED' | 'ALL';
 }
 
 export function isPropertyQueryType(queryType?: QueryType): boolean {
@@ -185,6 +196,23 @@ export function isPropertyQueryType(queryType?: QueryType): boolean {
 
 export function shouldShowLastObserved(queryType?: QueryType): boolean {
   return queryType === QueryType.PropertyAggregate || queryType === QueryType.PropertyValueHistory;
+}
+
+export function shouldShowOptionsRow(query: SitewiseQuery, showProp: boolean): boolean {
+  const shouldShowLastObservedSwitch =
+    shouldShowLastObserved(query.queryType) && !Boolean(query.propertyAlias) && showProp;
+  const shouldShowWithPropertyAlias =
+    // shouldn't show the row when querying associated assets with property alias, otherwise show it every time property alias is set
+    query.propertyAlias && !isListAssociatedAssetsQuery(query);
+  return !!(query.propertyId || shouldShowWithPropertyAlias || shouldShowLastObservedSwitch);
+}
+
+export function shouldShowL4eOptions(queryType?: QueryType): boolean {
+  return queryType === QueryType.PropertyValue || queryType === QueryType.PropertyValueHistory;
+}
+
+export function shouldShowQualityAndOrderComponent(queryType?: QueryType): boolean {
+  return queryType !== QueryType.PropertyValue;
 }
 
 // matches native sitewise API with capitals
@@ -213,11 +241,10 @@ export interface AssetInfo {
 /**
  * Metadata attached to DataFrame results
  */
-export interface SitewiseCustomMeta {
+export interface SitewiseCustomMetadata {
   nextToken?: string;
-
+  entryId?: string;
   resolution?: string;
-
   aggregates?: string[];
 }