+++ CHANGELOG.md
@@ -2,6 +2,148 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2.0.0
+
+- (Breaking change:) Migrate to plugin-ui from grafana/experimental in [#384](https://github.com/grafana/iot-sitewise-datasource/pull/384)
+- chore(deps): bump the all-go-dependencies group across 1 directory with 2 updates in [#413](https://github.com/grafana/iot-sitewise-datasource/pull/413)
+- chore(deps): bump the all-node-dependencies group across 1 directory with 20 updates in [#405](https://github.com/grafana/iot-sitewise-datasource/pull/ 405)
+- Bugfix: Filter out invalid queries in variable editor in [#399](https://github.com/grafana/iot-sitewise-datasource/pull/399)
+- Chore: update docker compose to support improved backend development in [#397](https://github.com/grafana/iot-sitewise-datasource/pull/397)
+- Bump the all-node-dependencies group with 7 updates in [#396](https://github.com/grafana/iot-sitewise-datasource/pull/396)
+- fix: list associated assets query fields in [#394](https://github.com/grafana/iot-sitewise-datasource/pull/394)
+- feat: support asset selection by external id in [#390](https://github.com/grafana/iot-sitewise-datasource/pull/390)
+
+## 1.25.2
+
+- Bump the all-node-dependencies group across 1 directory with 17 updates in [#389](https://github.com/grafana/iot-sitewise-datasource/pull/389)
+- Bump the all-go-dependencies group with 3 updates in [#388](https://github.com/grafana/iot-sitewise-datasource/pull/388)
+- Bump github.com/grafana/grafana-plugin-sdk-go from 0.260.3 to 0.261.0 in the all-go-dependencies group in [#385](https://github.com/grafana/iot-sitewise-datasource/pull/385)
+- Bump the all-node-dependencies group across 1 directory with 20 updates in [#382](https://github.com/grafana/iot-sitewise-datasource/pull/382)
+- Bump the all-go-dependencies group across 1 directory with 2 updates in [#379](https://github.com/grafana/iot-sitewise-datasource/pull/379)
+
+## 1.25.1
+
+- Chore: Update plugin.json keywords [#359](https://github.com/grafana/iot-sitewise-datasource/pull/359)
+- Bump the all-github-action-dependencies group with 3 updates [#372](https://github.com/grafana/iot-sitewise-datasource/pull/372)
+- Bump the all-go-dependencies group with 4 updates [#373](https://github.com/grafana/iot-sitewise-datasource/pull/373)
+- Bump the all-node-dependencies group with 30 updates [#374](https://github.com/grafana/iot-sitewise-datasource/pull/374)
+
+## 1.25.0
+
+- Fix: batch entries up to the limit specified in the API [#355](https://github.com/grafana/iot-sitewise-datasource/pull/355)
+
+## 1.24.0
+
+- fix: add check for nil for property value [#352](https://github.com/grafana/iot-sitewise-datasource/pull/352)
+- Fix golangci-lint errors [#353](https://github.com/grafana/iot-sitewise-datasource/pull/353)
+- fix: migrate asset id on the frontend in the query editor [#350](https://github.com/grafana/iot-sitewise-datasource/pull/350)
+- Remove "ANY" as a query quality option [#347](https://github.com/grafana/iot-sitewise-datasource/pull/347)
+- Chore: Rename datasource file [#344](https://github.com/grafana/iot-sitewise-datasource/pull/344)
+- Add precommit hook [#338](https://github.com/grafana/iot-sitewise-datasource/pull/338)
+- Remove unused fields from Get property value query editor [#343](https://github.com/grafana/iot-sitewise-datasource/pull/343)
+- Remove ANY option for interpolated property quality [#342](https://github.com/grafana/iot-sitewise-datasource/pull/342)
+
+## 1.23.0
+
+- Feat: add list time series to query editor by @ssjagad in [#336](https://github.com/grafana/iot-sitewise-datasource/pull/336)
+- Chore: update dependencies in [#337](https://github.com/grafana/iot-sitewise-datasource/pull/337)
+- Migrate to new form styling in config and query editors [#332](https://github.com/grafana/iot-sitewise-datasource/pull/332)
+
+## 1.22.1
+
+- Fix: use ReadAuthSettings to get authSettings in [#333](https://github.com/grafana/iot-sitewise-datasource/pull/333)
+
+## 1.22.0
+
+- Update to use GetSessionWithAuthSettings [#330](https://github.com/grafana/iot-sitewise-datasource/pull/330)
+
+## 1.21.0
+
+- Refactor Paginator [#313](https://github.com/grafana/iot-sitewise-datasource/pull/313)
+- Fix e2e tests [#316](https://github.com/grafana/iot-sitewise-datasource/pull/316)
+- Added Stalebot [#314](https://github.com/grafana/iot-sitewise-datasource/pull/314)
+- Added lint rule [#319](https://github.com/grafana/iot-sitewise-datasource/pull/319)
+- Add a frontend cache for relative time range queries [#318](https://github.com/grafana/iot-sitewise-datasource/pull/318)
+- Removed unsupported time ordering from interpolated query and fix issue with caching `aggregates` queries [#323](https://github.com/grafana/iot-sitewise-datasource/pull/323)
+
+## 1.20.0
+
+- Perf: Update batch api queries to request maximum number of dependencies in (#310)[https://github.com/grafana/iot-sitewise-datasource/pull/310]
+
+## 1.19.0
+
+- Fetch properties using both AssetId and AssetIds in (#307)[https://github.com/grafana/iot-sitewise-datasource/pull/307]
+- Migrate to CustomVariableSupport in (#304)[https://github.com/grafana/iot-sitewise-datasource/pull/304]
+
+## 1.18.0
+
+- Fix fetching asset properties in (#302)[https://github.com/grafana/iot-sitewise-datasource/pull/302]
+- Feature: L4E struct support in (#300)[https://github.com/grafana/iot-sitewise-datasource/pull/300]
+- Response processing: Add struct data type handling in (#297)[https://github.com/grafana/iot-sitewise-datasource/pull/297]
+- Query Editor: Improve menu placement for dropdowns in (#292)[https://github.com/grafana/iot-sitewise-datasource/pull/292]
+- Add keywords in (#291)[https://github.com/grafana/iot-sitewise-datasource/pull/291]
+- E2E: Add happy path playwright query tests in (#290)[https://github.com/grafana/iot-sitewise-datasource/pull/290]
+- Query Editor: Stop running queries on every change in (#274)[https://github.com/grafana/iot-sitewise-datasource/pull/274]
+
+## 1.17.0
+
+- Update grafana-aws-sdk to 0.21.0 and prepare 1.16.2 in (#282)[https://github.com/grafana/iot-sitewise-datasource/pull/282]
+- fix: clear selected property when assets are removed (#278)[https://github.com/grafana/iot-sitewise-datasource/pull/278]
+- E2E Tests (#285)[https://github.com/grafana/iot-sitewise-datasource/pull/285]
+- Add support for composite model properties (#279)[https://github.com/grafana/iot-sitewise-datasource/pull/279]
+- Use non-batch APIs at the edge (#281)[https://github.com/grafana/iot-sitewise-datasource/pull/281]
+
+## 1.16.1
+
+- Upgrade aws-sdk-go to v1.49.6 to have access to the `ExecuteQuery` API (#266)
+- Fix: Infer data type for disassociated streams for property value queries by alias (#275)
+
+## 1.16.0
+
+- Use query region to get client for queries (#258)
+- Feat: implement an "all" option for list associated assets query (#261)
+
+## 1.15.0
+
+- Support multiple assets for interpolated queries in (#256)[https://github.com/grafana/iot-sitewise-datasource/pull/256]
+
+## 1.14.0
+
+- Query and Config editors: Migrate to new form styling under feature toggle in (#244)[https://github.com/grafana/iot-sitewise-datasource/pull/244]
+
+## 1.13.0
+
+- Update dependencies and create-plugin configuration by @idastambuk
+  in https://github.com/grafana/iot-sitewise-datasource/pull/243
+- Property aggregate processing: Move out ErrorEntries processing from SuccessEntries block by @idastambuk
+  in https://github.com/grafana/iot-sitewise-datasource/pull/240
+- Bump go.opentelemetry.io/contrib/instrumentation/net/http/httptrace/otelhttptrace from 0.37.0 to 0.44.0 by @dependabot
+  in https://github.com/grafana/iot-sitewise-datasource/pull/241
+- Bump @babel/traverse from 7.17.10 to 7.23.2 by @dependabot
+  in https://github.com/grafana/iot-sitewise-datasource/pull/245
+- Bump loader-utils from 2.0.2 to 2.0.4 by @dependabot in https://github.com/grafana/iot-sitewise-datasource/pull/248
+- Bump semver from 5.7.1 to 5.7.2 by @dependabot in https://github.com/grafana/iot-sitewise-datasource/pull/247
+- Bump google.golang.org/grpc from 1.58.2 to 1.58.3 by @dependabot
+  in https://github.com/grafana/iot-sitewise-datasource/pull/246
+- Upgrade yaml package by @fridgepoet in https://github.com/grafana/iot-sitewise-datasource/pull/249
+- Upgrade underscore, debug dependencies by @fridgepoet in https://github.com/grafana/iot-sitewise-datasource/pull/252
+- Bump yaml from 2.2.1 to 2.3.4 by @dependabot in https://github.com/grafana/iot-sitewise-datasource/pull/253
+- Bump json5 from 2.2.1 to 2.2.3 by @dependabot in https://github.com/grafana/iot-sitewise-datasource/pull/254
+
+**Full Changelog**: https://github.com/grafana/iot-sitewise-datasource/compare/v1.12.1...v1.13.0
+
+## v1.12.1
+
+- Disassociated streams: Hash entryId to fix bug with property aliases longer than 64 characters in [#239](https://github.com/grafana/iot-sitewise-datasource/pull/239)
+
+## v1.12.0
+
+- Query by property alias: Add support for unassociated streams in [#231](https://github.com/grafana/iot-sitewise-datasource/pull/231)
+
+## v1.11.1
+
+- Revert "Replace deprecated setVariableQueryEditor with CustomVariableSupport" in [#229](https://github.com/grafana/iot-sitewise-datasource/pull/229)
+
 ## v1.11.0
 
 - Update backend grafana-aws-sdk to v0.19.1 to add `il-central-1` to the opt-in region list